extract out mark struct and transformer
